+++ src/api/mod.rs
@@ -1,0 +1,9 @@
+use crate::authentication::Authentication;
+
+pub mod edge_app;
+pub mod version;
+pub mod asset;
+
+pub struct Api {
+    pub authentication: Authentication
+}
